Minor improvements and refactoring on transactions-generator
