+++ transactions-generator/src/main/java/ba/klika/focus/reactive/transactions/transactionsgenerator/model/Transaction.java
@@ -1,11 +1,11 @@
 package ba.klika.focus.reactive.transactions.transactionsgenerator.model;
 
-import lombok.Data;
-import lombok.experimental.Accessors;
-
 import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
+
+import lombok.Data;
+import lombok.experimental.Accessors;
 
 @Data
 @Accessors(chain = true)
