+++ transactions-generator/src/main/java/ba/klika/focus/reactive/transactions/transactionsgenerator/util/TransactionGenerator.java
@@ -1,12 +1,7 @@
 package ba.klika.focus.reactive.transactions.transactionsgenerator.util;
 
-import ba.klika.focus.reactive.transactions.transactionsgenerator.model.Transaction;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.core.io.Resource;
-import org.springframework.stereotype.Component;
-import reactor.core.publisher.Flux;
-
 import java.io.BufferedReader;
+import java.io.IOException;
 import java.io.InputStreamReader;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
@@ -18,9 +13,15 @@ import java.util.Random;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
+import ba.klika.focus.reactive.transactions.transactionsgenerator.model.Transaction;
+import reactor.core.publisher.Flux;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.core.io.ClassPathResource;
+import org.springframework.core.io.Resource;
+import org.springframework.stereotype.Component;
+
 @Component
-public class TransactionsGenerator {
-    private static final String DATA_SOURCE = "data/MOCK_DATA.csv";
+public class TransactionGenerator {
     private static final Random random = new Random();
     private static final int FLUX_INTERVAL_MS = 1500;
     private static final long MIN_AMOUNT = 100;
@@ -28,23 +29,27 @@ public class TransactionsGenerator {
 
     private final List<String> ibanList;
 
-    public TransactionsGenerator() throws Exception {
+    @Value("${data.source.path}")
+    private String dataSource;
+
+    public TransactionGenerator() throws IOException {
         this.ibanList = getIbanListFromCsv();
     }
 
+
     public Flux<Transaction> getTransactions() {
         return Flux.interval(Duration.ofMillis(FLUX_INTERVAL_MS))
-            .map(this::generateTransaction)
-            .share();
+                .map(this::generateTransaction)
+                .share();
     }
 
     private Transaction generateTransaction(long i) {
         return new Transaction()
-            .setId(UUID.randomUUID())
-            .setSource(getIbanFromList())
-            .setDestination(getIbanFromList())
-            .setAmount(getRandomAmount())
-            .setCreatedAt(Instant.now());
+                .setId(UUID.randomUUID())
+                .setSource(getIbanFromList())
+                .setDestination(getIbanFromList())
+                .setAmount(getRandomAmount())
+                .setCreatedAt(Instant.now());
     }
 
     private String getIbanFromList() {
@@ -53,16 +58,16 @@ public class TransactionsGenerator {
     }
 
     private BigDecimal getRandomAmount() {
-        return new BigDecimal(MIN_AMOUNT + random.nextDouble() * (MAX_AMOUNT - MIN_AMOUNT))
-            .setScale(2, RoundingMode.HALF_UP);
+        return BigDecimal.valueOf(MIN_AMOUNT + random.nextDouble() * (MAX_AMOUNT - MIN_AMOUNT))
+                .setScale(2, RoundingMode.HALF_UP);
     }
 
-    private List<String> getIbanListFromCsv() throws Exception {
-        Resource sourceCsv = new ClassPathResource(DATA_SOURCE);
+    private List<String> getIbanListFromCsv() throws IOException {
+        Resource sourceCsv = new ClassPathResource(dataSource);
         BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(sourceCsv.getInputStream()));
 
         return bufferedReader.lines().skip(1)
-            .map(iban -> iban.replace(" ", ""))
-            .collect(Collectors.toList());
+                .map(iban -> iban.replace(" ", ""))
+                .collect(Collectors.toList());
     }
 }
